Fix multipath device discovery when UFN is enabled.
This currently returns an invalid path of `/dev/mapper/${WWID}`
when UFN is enabled leading to failures later on when we attempt to
use the device.

The output of `multipath -l ${path}` or `multipath -l ${wwid}`
should always list the correct device identifier to use with this
path as the first word on the first line.

The same change has been suggested for both Cinder [1] and Nova [2]
as they are also susceptible to this issue.

[1] https://review.openstack.org/#/c/170157/
[2] https://review.openstack.org/#/c/169873/

Change-Id: I4a2d0e6ba4522ebfa0c50be49f7039f6e4617ae8
Closes-Bug: 1401799