FC Eliminate the need to return devices list
This patch removes the need to build the list of devices found
when multipath was not being used.   At disconnect_volume time
for FC, we now use the device path to discover the actual device
information to remove the SCSI device from the system instead of
relying on the devices list, which isn't used by multipath anymore.

Change-Id: I5357c8cdf0976eee80120636f0a53f711db23016
Related-Bug: 1406161