Added ABCMeta class to the InitiatorConnector
This patch adds python ABCMeta class to the InitiatorConnector
for the purposes of defining a required contract that all Connector
objects must adhere to.  Currently, the existing required methods of
connect_volume, disconnect_volume are the 2 required methods that must
be implemented by the child Connector objects.  This will be helpful
for future features and forcing Connectors to adhere to the API
interface Contract.

Change-Id: Ie862ad7f358a70d57eddb2bce269230df3c3cdc7
Implements: blueprint os-brick-abc-connectors