Add new Connector APIs for path validation
The purpose of this patch is to add new APIs to os-brick Connector objects
to see how we can help validate the discovered volumes.  By adding a few
new APIs into the Connectors, we can ask the Connector objects to give us
a list of paths on the system that are associated with a volume.
Then after doing a disconnect, which happens during nova live migration
as well as normal volume detaches, we can fetch the list of all volumes, and
make sure the original volume paths don't exist in the list of all
volumes for that type.  Therefore verifying that when a volume
is removed, it doesn't leave any paths on the system behind.

This patch adds the following new APIs to the Connector objects:
 * get_volume_paths() - fetch a list of paths on the system for a volume.
 * get_search_path() - get the path on the host where the Connector looks for
                       volumes.
 * get_all_available_volumes() - Get the entire list of volumes in the search
                                 path for the given protocol type (iSCSI, FC)

Cinder Spec:
  https://review.openstack.org/#/c/204180/

Implements: blueprint brick-fetch-paths
Change-Id: Ibf73545d81797ffeedd7d4bf10c6d11c63fce923