FC discover existing devices for removal
Due to the way Nova live migration works, we can't
rely on having the device path that was discovered at
connect_volume time existing in the device_info.

This patch uses the existing connection_properties
to rediscover the existing volume paths on the system
at the time disconnect_volume is called.   This ensures
that whatever paths/devices are associated with at volume
at disconnect_volume time exists, they will be removed.

Change-Id: I699fcff5dee595caa54ca7e80f67966edd1370ef
Closes-Bug: 1406161