Add rootwrap filters
This patch adds os-bricks list of rootwrap filters for commands
that are needed to execute.   The filters are a self contained entire
list of expected filters that os-brick needs to run.  It's expected that
this filter file is added to any rootwrap enabled service that needs to use
os-brick.

Devstack associated patch: https://review.openstack.org/#/c/207677/

Partial-Bug: #1479842
UpgradeImpact:  Need to place the os-brick.filters file in service's
                rootwrap.d directory to enable filters.

Change-Id: I2b1e657b87c7b27548200a20b991f34c3413c24b