Use 'device' instead of 'volume_path'
If device is not found at the first try,
exception VolumeDeviceNotFound will be thrown.
The exception VolumeDeviceNotFound expects the key 'device',
however, 'volume_path' is passed in. As a result,
KeyError: 'device' exception will be thrown.

See the following definition in exception.py:
class VolumeDeviceNotFound(CinderException):
    message = _('Volume device not found at %(device)s.')

This bug is introduced in https://review.openstack.org/#/c/213389

Change-Id: I64b37486631c31569fb037064c2ec0fee3bd855c
Closes-Bug: #1492936