Multipath Device Action Being Parsed as Name
Whenever find_multipath_device() was being called on an action
it would include the action as part of the device name. This patch
creates a list of actions and checks that they aren't preceding
the device name before parsing the rest of the device information
in find_multipath_device().

This patch is based on the approach that was taken in Nova until
it was abandoned in favor of the change being made in os-brick
instead. The Change-Id for the Nova change:
I8f1f769bb15ec10054e2fd73c4cf621389014cf1

Change-Id: I12a10ba189e4987cdf07bebdadbac0d4587c01d1
Closes-Bug: 1466444