Fix test_connect_volume when skip is bypassed
If you run the unit tests with a live /dev/disk/by-path
on the system, the tests fall through the skipUnless
decorators.  This exposed some failures in the expected
commands, which were incorrect.   The expected commands
in _test_connect_volume had assumed the multipath set of
commands which included a rescan.

This patch removes the rescan expected tests, as those
are only run for multipath related attaches.

Closes-Bug: #1514611
Change-Id: If2e77af2a63c33df2f190f77b6ba4cea0bc31b31