os-brick add extend_volume API
This patch adds the new extend_volume API to the connector objects.
The purpose of this patch is to add the ability to notify the host
kernel when an attached volume has been resized.   This allows both
raw device paths and multipath devices to get new sizes on the fly
without detaching and then reattaching to the host.

Cinder Spec:
  https://review.openstack.org/#/c/243730/

To manually test:
  Use the brick_resize.py tool here:
  https://github.com/WaltHP/diediedie/blob/master/diediedie/brick_resize.py

  sample brick.conf to output debug logging for the brick_resize.py tool
  https://gist.github.com/WaltHP/e9b71e0ea1b1d9097762

  After a successful resize, find the virt domain name
  then run
  virsh blockresize <domain name> <path to raw device> <new size>
  for example,
  blockresize instance-00000007 /dev/disk/by-id/dm-uuid-mpath-350002ace1dda383d 3145728

Implements blueprint: brick-extend-attached-volume

Change-Id: I19858dd47afe3bdb3dd25c1dfead732a804cdee2