Fix iopsLimit parameter in ScaleIO connector
The parameter 'iops_limit' in the ScaleIO connector
is wrong. The correct parameter should be 'iopsLimit'.
Because of this error, the iops of the volume will stay
unlimited even if the user sets a limit on the iops.
This patch fixes this error.

Change-Id: I6b9212e9046bf3e2c9bbbde6354c298b359de3a6
Closes-Bug: #1521396
(cherry picked from commit 665ffea7307a9012f532a35c074667dc2963ad45)