Copy encryptors from Nova to os-brick
Currently, when creating an encrypted volume from an image, Cinder
writes raw data to the encrypted volume. When Cinder uploads an
encrypted volume to an image, it writes encrypted data to the image.
As a result, Nova cannot use these images or volumes.
To fix above problem, cinder needs to add encryptor attach/detach
layers.
As both Nova and Cinder needs to use the module, the fix is to
move it to os-brick.
It copies encryptors from Nova to os-brick, and keep all interfaces
unchanged except initialization.

Change-Id: I8044183ad02110c8b2468e20327d822c0437c772
Implements: blueprint improve-encrypted-volume
Related-bug: #1482464
Related-bug: #1465656