Remove multipath -l logic from ISCSI connector
This patch continues work on making the connect_volume methods
more efficient. Using multipath -l to find the available multipath
devices can take a lot of time listing all of the potential
devices. Previously, the FC connector had been modified to skip
the use of multipath -l.

This patch takes the FC code and makes it into a generic method to
be used in the ISCSI connector and the FC connector and updates tests.

I also fixed argument order in assertEquals() in the
test_connect_volume_with_multipath() function.

Closes-Bug: #1487169
Change-Id: If480df7c17a6f1c2ecebc0e00f5938d3056776fd