os-brick refactor get_connector_properties
This patch changes how the get_connector_properties works.  It migrates
the code for each specific connector type's connector_properties
(initiator) into the connector itself.   It also only calls each
connector that matches the platform and the os type.  This will allow
us to add Windows based connectors in the future.  So, when os-brick is
running on windows, it won't call any of the linux based connectors to
fetch the initiator information, as they won't work.

I'm not sure it's necessary to filter by platform as well, but it seemed
easy to do at the time.

Change-Id: I32290beea81d2a39b828fd8bf3ef805358c7f971