Replace _get_multipath_device_name with _discover_mpath_device
For ISCSIConnector, _get_multipath_device_name is used for discovery
of a multipath device. However, the functionality is duplicated with
InitiatorConnector's _discover_mpath_device.

_discover_mpath_device is more generic. So get_multipath_device_name
should be replaced.

This patch removes _get_multipath_device_name from ISCSIConnector and
replace it with _discover_mpath_device. And the related tests are removed.

Change-Id: I15e2cfe48468b52090b677a2b1c6175420a8c9ec
Close-Bug: #1550651