Fix multipath iSCSI encrypted volume attach failure
This is to copy fix of bug 1439869 from Nova to os-brick.

Currently iSCSI volume attachment fails if iscsi_use_multipath is
set to True. This is because the encryptor requests cryptsetup
to create the symlink to the LUKS device with the same name of
the device-mapper multipath device. To avoid the name collision,
this patch adds the 'crypt-' prefix to the symlink.

Change-Id: Ia001204df9b14f635ab998590e8add119c1aec23
Closes-Bug: #1439869