Detect if Fibre Channel support exists
This patch adds a simple check to see if the kernel
even supports Fibre Channel before calling systool -c fc_host -v
to find the active HBA's on the system.

The problem is that privsep can't quietly run systool without puking out
a bunch of errors to stderr.  We know that most systems don't have FC and
that's why we always wrapped out call to putils executing systool
with a try block.  putils allows callers to expect exceptions and then
handle the exception logging or not themselves.  privsep doesn't allow that.

Change-Id: Idd254303c9cca2b57d358de2fdbaf8abdd76b553