Windows remotefs: create mountpoints at the expected location
In order to support running inside a Scale-Out File Server cluster,
when 'mounting' a local share, we're creating a symlink to the local
export path without actually creating a SMB mapping.

The issue is that instead of creating a symlink using the hash of
the share UNC path, we're using the local export path instead. This
is undesired behavior, in which case the caller will not be able
to find the 'mountpoint'.

This change addresses this issue, ensuring that we're using the
mountpoint expected by the caller (e.g. the Cinder SMB driver)
when mounting SMB shares.

Closes-Bug: #1631351

Change-Id: I6bf909da559a745596f7d4c30d12e794b0b156de