Raise specific exception for an invalid protocol connector
Nova's libvirt driver attempts to load up a set of volume
drivers which use os-brick connectors on start of the
nova-compute service. If there is a missing or invalid
mapping, like ISER + s390x, it results in a ValueError and
nova-compute fails to start.

This change raises a specific exception, which is still a
ValueError for backward compatibility, so that Nova can
handle this situation gracefully and not crash.

Change-Id: I1d016c226416cf2b77d6e0393498905f56fd8cde
Partial-Bug: #1639239