Windows connectors: add device_scan_interval arg
The Hyper-V Nova driver will pass os-brick the device rescan
interval. This change indends to honor this for the Windows
connectors.

Other connectors use either a fixed rescan interval, either a
exponentially growing interval. If needed, this patch can be
updated so that other connectors may use this argument as well.

Change-Id: I3d6b4ed2ac27269241420ca03ec7075a977e86e8