Add Constraints support and missing bindep.txt dependencies
Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove os-brick.

Finally, it also adds libxml2 and libxslt to bindep.txt for use when
building lxml that is now required by oslo.vmware.

Change-Id: I3fbe4acbdf0281c91801ff7d09de6e9043915939