Encryptors: Fix compat with Nova encryptors for Ocata
The change 916cfaf "Introduce encryption provider constants"
intended to issue deprecation messages for encryptors that
are no longer going to be supported in Pike, and continue
loading the current os_brick encryptors instead.

However, since it removed the "'LuksEncryptor' in provider"
style checks, any config values that were not explicitly listed
in LEGACY_PROVIDER_CLASS_TO_FORMAT_MAP will not be translated.

This results in a configuration using a provider of
"nova.volume.encryptors.luks.LuksEncryptor" to attempt to
import this module directly into Cinder when previously it
would use the os_brick encryptor.  This is generally wrong
and causes a number of problems.

Closes-Bug: #1658755

Change-Id: I3ec6e3fe919bc03d158da04a18fb8b651002ed52