Change log level on _get_hba_channel_scsi_target
On the FC initiator, when we are doing a scan we try to narrow the scope
of the scan as much as possible, for this purpose we try to get the HBA
channel and SCSI targets for an HBA.

Unfortunately the mechanism we have implemented to detect these two
parameters is not universal and will only work on FC targets that
implement a single WWN.

This is not a big problem as the code is capable of operating normally
without those 2 parameters, it's just that the filtering on the scan is
not as precise as it could be.

So the issue is that we are logging as an error when we cannot get that
information, which is clearly misleading to anybody reading the logs.

This patch changes the log level from error to debug.

TrivialFix

Change-Id: Ieec4868b9da0d0af5b951fb01257a742c440aeee