Add attribute 'name' to class RBDVolume.
If we create backup for volume  in-use, volume
driver will clone a temporary volume, and create
a snapshot based on the temporary volume, then
create backup for the snapshot. So, the src_name
passed to method _rbd_diff_transfer should be the
name of the temporary volume instead of the original
volume's name. In order to get this name, we need
add self.name to class os_brick.initiator.linuxrbd.
RBDVolume, that's because ceph provides rbd.so to
cinder, but we can't get volume name from rbd.Image
directly.

Co-Authored-By: Ivan Kolodyazhny <e0ne@e0ne.info>
Change-Id: Ic2732e985b75ae710da3814c2d5b2253ced74534
Closes-Bug: #1663504