Retry multipath flush when map is in use
When flushing a multipath we are ignoring map in use transient error, so
we log a warning that the flush has not been successful and that we have
received an exit code 1 and we continue to remove the individual paths.

This error is usually transient and a simple retry will succeed in
flushing the multipath.

Failure to retry will leave an empty multipath in our system.

Closes-Bug: #1663936
Depends-On: I31310dc9b3834e0bf16618db3e473f17eab2b610
Change-Id: I710792bd707ad933ef60d11d25f530dddfb6fb2f
(cherry picked from commit 32c837dacb38910e9191b1a3f18283b3c52fb2b2)