Retry multipath flush when map is in use
When flushing a multipath we are ignoring map in use transient error, so
we log a warning that the flush has not been successful and that we have
received an exit code 1 and we continue to remove the individual paths.

This error is usually transient and a simple retry will succeed in
flushing the multipath.

Failure to retry will leave an empty multipath in our system.

Closes-Bug: #1663936
Change-Id: I710792bd707ad933ef60d11d25f530dddfb6fb2f