RBD: consider a custom keyring in connection info
If a 'keyring' key is found in the connection info passed to
connect_volume() use its value as the path to the keyring instead of the
default location (/etc/ceph/<cluster>.client.<user>.keyring).

This allows services such as cinder's RBD and Ceph backup drivers to
make use of a custom keyring path that an admin has defined.

Change-Id: Ib1230d3e40f56371567e1aead40db59667bad295
Closes-bug: #1668304