iSCSI multipath: improve logging on connect
The new iSCSI multipath code uses multithreading to do the connections,
but since the context is a thread local variable it will not be present
in those threads so we will not see the request id or other information
from the context in those thread's log entries.

This is a small inconvenience when looking at the logs if we try to
filter by the request id,.

To resolve this we introduce a new Thread class that inherits the
context from the calling thread, this way any log entries will have the
right context and present expected information.

Change-Id: I27624a59738809d30531c0c4f37308f897dd9654