Fix manual scan for discovery type backends
When we added open-iscsi manual scan support we made it so that it
wouldn't modify the scan mode of already existing nodes, and this is a
problem for backends that use discovery, since the discovery will add
the nodes with the default automatic scan mode.

This patch makes sure that we always try to change the mode to manual
scans and only reports scans to be automatic if the initiator doesn't
support this feature.

TrivialFix

Change-Id: Id9e2f539d5273a20038399d9af6017652a4f987f