Fix iSCSI cleanup fix on discovery backends
On Change-Id Iada5d4fbeb07aeaf3afb953a289b6b89778c382c we tried to fix
an issue with the multipath detach of backends that used discovery on
attach, but contrary to the commit message and the docstrings it didn't
look for ip,port in the discoverydb but ip:port instead, which meant
that it would never find what it was looking for.

This patch fixes that fix to make it search for the right regex.

TrivialFix
Closes-Bug: #1699061

Change-Id: Ibfa1a78a555e984c662f668677451f5a3ed55602