Revert "Don't use ignoreskipactivation for thin LVM"
Commit b82ab0d1a078d435bfe85c8aa9ed0805f26ca63a wrongly assumed that
the "ignoreskipactivation" flag is not needed when using thin LVM
provisioning. This is wrong.
Using thin LVM leads now to an error when the content will be copied:

ERROR cinder.volume.manager Command: sudo cinder-rootwrap /etc/cinder/
      rootwrap.conf dd if=/dev/mapper/cinder--volumes-_snapshot--UUID
      of=/dev/mapper/cinder--volumes-volume--UUID
      count=1073741824 bs=1M iflag=count_bytes conv=sparse

The reason is that the the snapshot is not activated and due to that,
no device node is available under /dev/mapper .

This reverts commit b82ab0d1a078d435bfe85c8aa9ed0805f26ca63a and makes
thin provisioning for LVM working again.

Change-Id: I062fd7d7120081c79c714d8177104cdd95522d2d
Depends-On: I788b297af92e11063b18c05a415b6f4d4d4bcd2c
Closes-Bug: #1702322
Related-Bug: #1701545