Fix _remove_scsi_symlinks_no_links test
In test_linuxscsi's `_remove_scsi_symlinks_no_links` test we are passing
the list of devices to the method including the /dev prefix, when it
shouldn't have it.

The test result is the same, but it is incorrect to do so as it would be
missing the intent of the test case.

TrivialFix

Change-Id: If0925a5aa3b22248027e6dbae7954e573a1476b8