Windows: fix connectors 'disconnect_volume' signature
The Windows connectors do not include the 'device_info' argument
in the 'disconnect_volume' method.

This wasn't an issue for Nova as the Hyper-V driver doesn't pass
this argument, yet Cinder does pass it, which will lead to an
exception.

This change adds the missing argument, defaulting to "None" so that
it won't break the Nova Hyper-V driver.

Change-Id: Iaf8b57c97afc393862a61a7338a35f008b65c8c1
Closes-Bug: #1738390