Handle multiple errors in multipath -l parsing
The current code in linuxscsi.py handles the case where there is one
error in the output of "multipath -l", but when there are multiple
errors there is an empty line between error messages. The code filters
out the error lines but not the empty lines, and so it then looks for
the device name in an empty line and so it fails to parse it.

Closes-Bug: #1778468
Change-Id: I7a5bdf285a74e045ce7d11df5e8c0921d3447106
(cherry picked from commit 9b729ef0e7461f5ca2d80b7f991aeacf547a1819)