Fix multipath disconnect with path failure
Under certain conditions detaching a multipath device may result on
failure when flushing one of the individual paths, but the disconnect
should have succeeded, because there were other paths available to flush
all the data.

OS-Brick is currently following standard recommended disconnect
mechanism for multipath devices:

- Release all device holders
- Flush multipath
- Flush single paths
- Delete single devices

The problem is that this procedure does an innecessary step, flushing
individual single paths, that may result in an error.

Originally it was thought that the individual flushes were necessary to
prevent data loss, but upon further study of the multipath-tools and the
device-mapper code it was discovered that this is not really the case.

After the multipath flushing has been completed we can be sure that the
data has been successfully sent and acknowledge by the device.

Closes-Bug: #1785669
Change-Id: I10f7fea2d69d5d9011f0d5486863a8d9d8a9696e