Improve iSCSI device detection speed
Current iSCSI device detection checks for the presence of devices based
on the scan time, so it checks for presence, sleeps, scans, and checks
again.  This means that if the device becomes available while we sleep
to send another scan we won't detect the device.

This patch changes this, making the searching and the rescanning
independent operations operating at different cadences.

Checking for the device will happen every seconds, and the rescans will
happen after 4, 9, and 16 seconds.

Change-Id: I716a3ea8583e289819cc37b6b5dd9730dd59406b