Improve detection of multipathd running
There is a case where `multipathd show status` should fail, yet it
doesn't exit with an error code. Instead it returns 0. Error is printed
to stdout as "error receiving packet".

This means that enforce will not work as expected.

The multipathd error got fixed here:
https://www.redhat.com/archives/dm-devel/2015-March/msg00136.html

But that patch is missing in some packages, so this patch adds a
workaround to detect the error base on the stdout of the multipathd show
status command.

Closes-Bug: #1789699
Change-Id: I1cb29782541e3fe53b0b1744ab36f56464cd2135