Optimize FC device checking
Checks for fibre channel device paths will only fail if there are
no FC HBAs since the actual paths are just calculated based on the
HBA information, whether they exist or not. This moves the check
earlier in the process to avoid that calculation if we don't have
any HBA information.

Change-Id: Ib8bfea2c9603335e93b1407605f0e59d3c29f622